Update using-cfn-waitcondition.md
The WaitCondition doesn't really "pause the stack", nor does it help define when a resource is "really created" (that's the CreationPolicy).  Instead, it allows a dependency to be setup between resources where a specified number of signals must be received before the dependent resource can begin being created.  Resources unrelated to the wait condition are free to be created.