ReceiptRuleSet has Return Values provided with Ref
While helping a user of troposphere come up with a way to configure SES
`ReceiptRule`s/`ReceiptRuleSet`s I noticed that for some reason the docs
didn't list any Return Values for `ReceiptRuleSet`, which would make it
very difficult to configure in Cloudformation, and seemed like an
oversight.  I went ahead and tried using `Ref` on a `ReceiptRuleSet` to
provide an output of the physical resource created, and it appears to
work.

This is the example YAML I used:

```
AWSTemplateFormatVersion: 2010-09-09
Description: 'AWS SES ReceiptRuleSet Sample Template'
Resources:
  ReceiptRuleSet:
    Type: 'AWS::SES::ReceiptRuleSet'
Outputs:
  ReceiptRuleSetRef:
    Value: !Ref ReceiptRuleSet
```

And here's the resulting output:

https://www.dropbox.com/s/hwedbwyn3gwezrw/Screenshot%202018-05-19%2007.46.29.png?dl=0