Update aws-properties-ec2-instance-launchtemplatespecification.md
I have tested this with CFN that if I give Launch Template Name for LaunchTemplateId, I get an error saying, I must give Launch Template ID for the LaunchTemplateId. 
And same for the LaunchTemplateId, It needs Launch Template ID and not the Launch Template Name.
Previously the docs was suggesting that, you can give either Launch Template Name or Launch Template ID for both the properties, hence corrected the same.