Make documentation clearer around dependency
This has cost us quite a bit of time.
The documentation as it is states that there is a dependency from the delivery channel to the Configuration Recorder. Once we added a `DependsOn` to the delivery channel resource the creation of the Configuration Recorder was stuck and eventually timed out because there was no delivery channel, so the recorder wouldn't start.
The documentation must be clearer that this sentence does not mean to add a `DependsOn` to the resource.