Update aws-resource-ec2-transitgatewayattachment.md
When I first read the returned value of the Ref function for this resource it made me come to the conclusion I needed to get the attachment ID in a different way. I actually created a CFN custom resource to do that.

I came back to this today and felt that it didn't make sense to return the name so I tested it. I have a cfn stack that creates multiple `AWS::EC2::TransitGatewayAttachment`'s.

The following does return the attachment ID not the name.

```
Outputs:
  LegacyDefaultVpcAttachmentId:
    Description: The ID of the attachment.
    Value: !Ref LegacyDefaultVpcAttachment
```

Please advise.