Correct `Parameters` property, add info and example
The previous value for *Type* that was `JSON` is incorrect.  This can lead to quite a bit of wasted time for users trying to craft a proper value for the `Parameters` property. This file change adds a few clarifying statements about the value for `Parameters`, and adds a YAML example of the syntax for the sub-properties of the `Parameters` property