More fully explain the relationship between EvaluationPeriods and DatapointsToAlarm
The existing docs are somewhat incomplete in that they
allude to the creation of "M out of N" alarms under
`DatapointsToAlarm`, and to "N consecutive" alarms under
`EvaluationPeriods`, but don't refer to the other related
category.

This PR is intended to clarify that by explaining that if
`DatapointsToAlarm` is unspecified, it defaults to matching
`EvaluationPeriods` (whence the "N consecutive" alarm
behavior).