Clustered vs Standalone Outputs
In the 'Return Values' section, there was no mention of which return values could be expected with what cluster mode. Cluster Mode Enabled, for example, only had the configuration endpoint, as where the Cluster Mode Disabled includes the primary and reader endpoints. The naming of the outputs doesn't make it immediately clear to a new user which outputs to expect, so these headers I think should give some quick direction.