Fix typo of MySQL to PostgreSQL in examples
In the example "Creating an Amazon Aurora DB cluster that exports logs to Amazon CloudWatch Logs" an Amazon Aurora PostgreSQL DB cluster cloudformation template is presented. In both the respective JSON and YAML examples the parameter descriptions speak about MySQL and not PostgreSQL.

This commit fixes this typo and includes PostgreSQL within the parameter descriptions.

See:
- https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-rds-dbcluster.html or
- https://github.com/awsdocs/aws-cloudformation-user-guide/blob/main/doc_source/aws-resource-rds-dbcluster.md#creating-an-amazon-aurora-db-cluster-that-exports-logs-to-amazon-cloudwatch-logs