Removed division column on RegularColumns.
As division column is created in ClusteringKeyColumns we cant create the division in RegularColumns if we run CFT with division being in both columns we get error and stack will destroy automatically. so to achieve the output we need to remove division in RegularColumns. So, I have updated the same please verify and approve. Thank you.