ECS TaskDefinition ResourceRequirement allows InferenceAccelerator
*Description of changes:*

The summary of this page says that a GPU is the only supported ResourceRequirement. However, the properties section states that `InferenceAccelerator` is also a supported value. This change updates the summary to reflect that difference.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.