Update aws-resource-ec2-vpc.md
The dedicated value in InstanceTenancy caused me a few hours of delay after causing a very vague error message in a cloudformation system I was starting to build. I finally found this solution https://stackoverflow.com/questions/45691826/aws-launch-configuration-error-the-requested-configuration-is-currently-not-sup and back traced the problem to this piece of documentation.  I think the most common reason someone would copy this code is a new build and unlikely that a newb would want dedicated tenancy. I think it would be best to set this flag to default so others don't get sent down this same rabbit hole.