Example : Continuous JSON-based anomaly detector
This adds JSON and YAML examples demonstrating a CloudFormation template which will create an `AWS::LookoutMetrics::AnomalyDetector` for continuous data and historical back-test data.

In the example the creation of an S3-based metric source for 5 minute aggregations is demonstrated. It includes an example of how to properly utilize the AWS date format characters to specify a path for continuous and historical data along with dimension selection and timestamp-field formatting examples for each JSON record in a continuous sample file.