AWS::Logs::ResourcePolicy GovCloud
Using this resource in GovCloud results in the following. 
`Template format error: Unrecognized resource types: [AWS::Logs::ResourcePolicy]`

Can we please add a disclaimer indicating this in the ACTUAL reference material?

If we have to change the format of it, or it needs to be rewritten, I don't care. But this is ridiculous to not have this information available. 

Related: 
- https://github.com/aws-cloudformation/cloudformation-coverage-roadmap/issues/351
- https://github.com/aws/aws-cdk/issues/5343#issuecomment-1207030757