Striving for clarity and simplicity
Removing a non-contradictory "but" statement. Perhaps nitpicking but the part Im removing sounds off to me since the first part is not opposite, in my mind its like saying "the day is nice but the sky is pretty". The part about the process being the same doesnt seem to match the subject or context of the first part of the sentence (before the comma), which process is the same?